file path changes

+++ ProEstimator.Admin/Controllers/EmailReportController.cs
@@ -81,15 +81,18 @@ namespace ProEstimator.Admin.Controllers
                     Directory.CreateDirectory(adminFolder);
                 }
 
-                if (System.IO.File.Exists(diskPath))
+                if(Path.GetFullPath(diskPath).StartsWith(adminFolder, StringComparison.OrdinalIgnoreCase))
                 {
-                    System.IO.File.Delete(diskPath);
+                    if (System.IO.File.Exists(diskPath))
+                    {
+                        System.IO.File.Delete(diskPath);
+                    }
+
+                    spreadsheetWriter.WriteSpreadshet(table, diskPath);
+
+                    byte[] fileBytes = System.IO.File.ReadAllBytes(diskPath);
+                    return File(fileBytes, System.Net.Mime.MediaTypeNames.Application.Octet, "EmailReport_" + startDateName + "_" + endDateName + ".xlsx");
                 }
-
-                spreadsheetWriter.WriteSpreadshet(table, diskPath);
-
-                byte[] fileBytes = System.IO.File.ReadAllBytes(diskPath);
-                return File(fileBytes, System.Net.Mime.MediaTypeNames.Application.Octet, "EmailReport_" + startDateName + "_" + endDateName + ".xlsx");
             }
             catch (Exception ex)
             {
