+++ ProEstimator.Admin/Controllers/PdrReportController.cs
@@ -75,15 +75,18 @@ namespace ProEstimator.Admin.Controllers
                         Directory.CreateDirectory(adminFolder);
                     }
 
-                    if (System.IO.File.Exists(diskPath))
+                    if(Path.GetFullPath(diskPath).StartsWith(adminFolder, StringComparison.OrdinalIgnoreCase))
                     {
-                        System.IO.File.Delete(diskPath);
+                        if (System.IO.File.Exists(diskPath))
+                        {
+                            System.IO.File.Delete(diskPath);
+                        }
+
+                        spreadsheetWriter.WriteSpreadshet(table, diskPath);
+
+                        byte[] fileBytes = System.IO.File.ReadAllBytes(diskPath);
+                        return File(fileBytes, System.Net.Mime.MediaTypeNames.Application.Octet, "PdrReport_" + startDateName + "_" + endDateName + ".xlsx");
                     }
-
-                    spreadsheetWriter.WriteSpreadshet(table, diskPath);
-
-                    byte[] fileBytes = System.IO.File.ReadAllBytes(diskPath);
-                    return File(fileBytes, System.Net.Mime.MediaTypeNames.Application.Octet, "PdrReport_" + startDateName + "_" + endDateName + ".xlsx");
                 }
                 catch (Exception ex)
                 {
